sched: init this_load to max in find_idlest_group
Date	Fri, 18 Oct 2013 13:52:23 +0200

Init this_load to max value instead of 0 in find_idlest_group.
If the local group is skipped because it doesn't have allowed CPUs, this_load
stays to 0,  no idlest group will be returned and the selected CPU will be a
not allowed one (which will be replaced in select_fallback_rq by a random
one). With a default value set to max, we will use the idlest group even if we
skip the local_group.

Change-Id: I4af96b66c1d22b23b9d3bd4433054f3e210dc1fc
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Paul Reioux <reioux@gmail.com>