sched: differantiate idle cpu
Date	Fri, 18 Oct 2013 13:52:26 +0200

The cost for waking up of a core varies according to its current idle state.
This includes C-state and intermediate state when some sync between cores is
required to reach a deep C-state.
Waking up a CPU in a deep C-state for running a short task is not efficient
from both a power and a performance point of view. We should take into account
the wake up latency of an idle CPU when the scheduler looks for the best CPU
to use for a waking task.
The wake up latency of a CPU is computed into a load that can be directly
compared with task load and other CPUs load.

Change-Id: Ic105205029c6534e72dbfe725d87e55952c3aa9d
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Paul Reioux <reioux@gmail.com>